Modified the insertion of the CRC into the buffer.
Fixed a bug that was occuring on some microcontrollers where the crc would be placed in the wrong place due to casting the buffer to a word from a char array.
Certain architectures were not allowing the word to be placed at a non-word memory boundary and the crc would over write the previous byte.